partial edit to test_ex

+++ tests/functional_tests/ffront_tests/test_execution.py
@@ -723,14 +723,14 @@ def test_conditional(fieldview_backend):
 
 def test_astype(fieldview_backend):
     size = 10
-    b = np_as_located_field(IDim)(np.ones(size))
+    b = np_as_located_field(IDim)(np.ones((size), dtype=np.float64))
     c = np_as_located_field(IDim)(np.ones((size,), dtype=bool))
     out_bool = np_as_located_field(IDim)(np.zeros((size,), dtype=bool))
-    out_int = np_as_located_field(IDim)(np.zeros((size,), dtype=int))
+    out_int = np_as_located_field(IDim)(np.zeros((size,), dtype=np.int64))
 
     @field_operator(backend=fieldview_backend)
-    def astype_fieldop_int(b: Field[[IDim], float64]) -> Field[[IDim], int]:
-        d = astype(b, int)
+    def astype_fieldop_int(b: Field[[IDim], float64]) -> Field[[IDim], int64]:
+        d = astype(b, int64)
         return d
 
     astype_fieldop_int(b, out=out_int, offset_provider={})
